Move texture import settings to OnPreprocessTexture.
Improves performance of texture importing over existing double-import
implementation. Verify this benefit by comparing the performance of
CopyClientData in this and the previous commit.

Necessary to be able to create a method in the next commit to be able to
save a texture without importing it in the middle of the function.

Being able to save a texture without immediately importing it is
necessary in order to batch texture imports.

Batching texture imports is necessary to increase the performance of
texture importing, including but not necessarily limited to the
performance of importing atlas textures generated from .spr files.

The performance increase of asset import batching is a property of
avoiding the cost of restarting Unity's asset importer over and over.

Batching texture imports is probably also necessary to support any
asset generation multi-threading desired to be done in the future for
further performance increases.

Note that importing cannot be multi-threaded, as it is controlled by
Unity, but generating assets to be imported can be multi-threaded, as it
is controlled by you.